created saveChatButton component only for logged in users
