+++ src/app/chatbot/page.tsx
@@ -8,6 +8,7 @@ import { Button, Form, OverlayTrigger, Popover } from "react-bootstrap";
 import { useSession } from "next-auth/react";
 import { FaRobot } from "react-icons/fa";
 import SidebarChat from "@/components/SidebarChat";
+import SaveChatButton from "@/components/SaveChatButton";
 
 export type ChatMessage = {
   role: string;
@@ -137,34 +138,6 @@ function NormalChat() {
     } catch (error) {
       console.log("error");
     }
-  };
-
-  const handleSaveChat = async (
-    e: React.MouseEvent<HTMLButtonElement, MouseEvent>
-  ) => {
-    e.preventDefault();
-
-    const myHeaders = new Headers();
-    myHeaders.append("Content-Type", "application/json");
-
-    const raw = JSON.stringify({
-      user_id: data?.user?.id,
-      messages: messages,
-    });
-
-    const requestOptions = {
-      method: "POST",
-      headers: myHeaders,
-      body: raw,
-    };
-
-    const response = await fetch(
-      "http://localhost:3000/api/chats",
-      requestOptions
-    );
-
-    const result = await response.json();
-    console.log(result);
   };
 
   function handleSelectionStart() {
@@ -333,11 +306,7 @@ function NormalChat() {
             ) : (
               <Button disabled>Clear</Button>
             )}
-            {messages.length > 1 ? (
-              <Button onClick={handleSaveChat}>Save</Button>
-            ) : (
-              <Button disabled>Save</Button>
-            )}
+            {data?.user ? <SaveChatButton messages={messages} /> : ""}
           </div>
         </Form>
       </div>
